pyproject.toml
[tool.poetry]
name = "Sleuth"
version = "2.1.0"
description = "Utility which watches the Fortnite News Feed for changes then shares it on Twitter."
authors = ["EthanC <EthanC@github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.11.1"
python-twitter = "^3.5"
coloredlogs = "^10.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
